blah blah blah

+++ av1/encoder/aq_rdo.h
@@ -20,7 +20,7 @@ extern "C" {
 
 void av1_rdo_aq_frame_setup(AV1_COMP *cpi);
 int av1_rdo_aq_seg_rate(AV1_COMMON *cm, MACROBLOCKD *xd, int segment_id, BLOCK_SIZE bs);
-int av1_rdo_aq_dist_scale(AV1_COMP *cpi, MACROBLOCK *x, BLOCK_SIZE bs);
+double av1_rdo_aq_dist_scale(AV1_COMP *cpi, MACROBLOCK *x, BLOCK_SIZE bs);
 int av1_rdo_aq_select_segment(AV1_COMP *cpi, MACROBLOCK *mb, BLOCK_SIZE bs);
 
 #ifdef __cplusplus
